refactor: fix bus list model style
